first commit
This banking record system project is developed without the use of the graphics component. It is more of a database project
in C++ and is built using the language’s file handling mechanism. Here, in this project, the user can choose the option they want to operate. The user can add the record to file, show the record or search any records from a file, can update the record, and also delete the record from the file as per your need.

The system asks users to enter their choice. The user can add the record with the account number, first name, last name, and the balance. Also, you can view or search the record details of all the stored records. You can update or delete the record with the help of the record numbers. Thus, this application works smoothly without any bug.

Features:
> Developed using the C++ language.
> Easy to operate and understandable.
> Based on CRUD operation